Conform to RFC 2253 / RFC 4514 for subject names.
Previously, we used the openssl defaults when printing subject names, but upon further examination, the RFC 2253/RFC 4514 format is in much more common usage, and are the standards used by LDAP and other DN systems.  This is by far the better default without exposing the myriad options for how to display subject names.